@@ -3,11 +3,12 @@ name = "soap-jax"
 version = "0.1.0"
 description = "SOAP Optimizer implemented in JAX"
 readme = "README.md"
-requires-python = ">=3.9"
+requires-python = ">=3.11"
 dependencies = [
     "jax",
     "jaxtyping",
-    "optax",
+    "optax>=0.2.6",
+    "flax>=0.12.0"
 ]
 
 [build-system]
@@ -17,7 +18,7 @@ build-backend = "hatchling.build"
 [tool.ruff]
 line-length = 120
 indent-width = 4
-target-version = "py39"
+target-version = "py311"
 
 [tool.ruff.lint]
 select = ["E", "F", "B", "SIM", "I", "FURB"]
